final-fixes

+++ ft_memmove.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void				*ft_memmove(void *dst, const void *src, unsigned int n)
+void				*ft_memmove(void *dst, const void *src, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*d;
 	unsigned char	*s;
